requirements.txt
google-cloud-bigquery
bs4
requests
pandas
fpdf2
matplotlib
cairosvg
selenium
